@@ -1,72 +1,77 @@
 {
   "name": "menatwork/contao-multicolumnwizard-bundle",
-  "type": "contao-bundle",
   "description": "MultiColumWizard for Contao OpenSource CMS",
+  "license": "LGPL-3.0-or-later",
+  "type": "contao-bundle",
   "keywords": [
     "contao",
     "widget",
     "wizard"
   ],
-  "homepage": "http://www.men-at-work.de",
-  "license": "LGPL-3.0-or-later",
   "authors": [
     {
       "name": "Stefan Heimes",
       "email": "stefan_heimes@hotmail.com",
-      "homepage": "http://www.men-at-work.de",
+      "homepage": "https://www.men-at-work.de",
       "role": "Developer"
     },
     {
       "name": "Andreas Isaak",
       "email": "contao@men-at-work.de",
-      "homepage": "http://www.men-at-work.de",
+      "homepage": "https://www.men-at-work.de",
       "role": "Developer"
     },
     {
       "name": "David Maack",
       "email": "contao@men-at-work.de",
-      "homepage": "http://www.men-at-work.de",
+      "homepage": "https://www.men-at-work.de",
       "role": "Developer"
     },
     {
       "name": "Andreas Schempp",
       "email": "andreas.schempp@terminal42.ch",
-      "homepage": "http://www.terminal42.ch",
+      "homepage": "https://www.terminal42.ch",
       "role": "Developer"
     },
     {
       "name": "Yanick Witschi",
       "email": "yanick.witschi@terminal42.ch",
-      "homepage": "http://www.terminal42.ch",
+      "homepage": "https://www.terminal42.ch",
       "role": "Developer"
     },
     {
       "name": "Tristan Lins",
-      "email": "tristan.lins@bit3.de",
-      "homepage": "http://bit3.de",
+      "homepage": "https://github.com/tristanlins",
       "role": "Developer"
     }
   ],
+  "homepage": "https://www.men-at-work.de",
   "support": {
     "issues": "https://github.com/menatwork/MultiColumnWizard/issues",
-    "wiki": "http://de.contaowiki.org/MultiColumnWizard",
+    "wiki": "https://de.contaowiki.org/MultiColumnWizard",
     "irc": "irc://irc.freenode.org/contao.dev",
     "source": "https://github.com/menatwork/MultiColumnWizard"
   },
   "require": {
-    "php": "^5.6 || ^7.0 || ^8.0",
-    "contao/core-bundle": "^4.9",
-    "symfony/config": "^4.4 || ^5.1",
-    "symfony/console": "^4.4 || ^5.1",
-    "symfony/dependency-injection": "^4.4 || ^5.1",
-    "symfony/event-dispatcher": "^4.4 || ^5.1",
-    "symfony/http-foundation": "^4.4 || ^5.1",
-    "symfony/http-kernel": "^4.4 || ^5.1",
-    "symfony/translation": "^4.4 || ^5.1"
+    "php": "^8.1",
+    "ext-json": "*",
+    "contao/core-bundle": "^4.13 || ^5.0",
+    "symfony/config": "^5.4 || ^6.0",
+    "symfony/console": "^5.4 || ^6.0",
+    "symfony/dependency-injection": "^5.4 || ^6.0",
+    "symfony/event-dispatcher": "^5.4 || ^6.0",
+    "symfony/http-foundation": "^5.4 || ^6.0",
+    "symfony/http-kernel": "^5.4 || ^6.0",
+    "symfony/translation": "^5.4 || ^6.0"
   },
   "require-dev": {
-    "phpcq/all-tasks": "~1.2",
-    "contao/manager-bundle": "^4.4"
+    "contao/manager-plugin": "^2.1",
+    "phpcq/runner-bootstrap": "^1.0@dev",
+    "roave/security-advisories": "dev-latest"
+  },
+  "replace": {
+    "contao-legacy/multicolumnwizard": "*",
+    "menatwork/contao-multicolumnwizard": ">=3.3.4 <4.0"
   },
   "autoload": {
     "psr-4": {
