@@ -22,7 +22,9 @@
 
 namespace MenAtWork\MultiColumnWizardBundle\EventListener\Mcw;
 
+use Contao\Backend;
 use Contao\BackendTemplate;
+use Contao\System;
 use MenAtWork\MultiColumnWizardBundle\Event\GetTinyMceStringEvent;
 
 /**
@@ -49,7 +51,7 @@ class TinyMce
         $fileBrowserTypes = array();
         // Since we don't know if this is the right call for other versions of contao
         // we won't use dependencies injection.
-        $pickerBuilder = \System::getContainer()->get('contao.picker.builder');
+        $pickerBuilder = System::getContainer()->get('contao.picker.builder');
 
         foreach (array('file' => 'image', 'link' => 'file') as $context => $fileBrowserType) {
             if ($pickerBuilder->supportsContext($context)) {
@@ -65,7 +67,7 @@ class TinyMce
         $objTemplate->source           = $table . '.' . $fieldId;
 
         // Deprecated since Contao 4.0, to be removed in Contao 5.0
-        $objTemplate->language = \Backend::getTinyMceLanguage();
+        $objTemplate->language = Backend::getTinyMceLanguage();
 
         $event->setTinyMce($objTemplate->parse());
     }
